improvements/fixes
