+++ src/lib/minions/decorators/index.ts
@@ -1,2 +1,3 @@
 export { default as minionNotBusy } from './minionNotBusy';
 export { default as requiresMinion } from './requiresMinion';
+export { default as ironsCantUse } from './ironsCantUse';
